Extend the protocol to be able to do asynchronous wifi scan
This change allows printer to ask esp to start scanning for available
networks. The results are stored on ESP side to preserve memory on the
printer. The results can be retrieved from the ESP over uart and the ESP
periodically updates the printer about number of found access points.

BFW-5387