Use common files with printer FW
Instead of specifing common data structures and constants twice, we
now keep common files with printer FW. This should increase the code
readability.

Right now the code is copied manually from the printer repo, but if we
ever want to move the esp firmware into the main repo we can then make
it a static library and recompile it for printer FW and esp FW alike.

On esp8266 we can't use all the features of c++20, so we don't have a
std::span. To make it easier to write the common code, we include part
of Microsoft's GSL library, that implements span.

BFW-5659